Arreglado el asuntito con los try catch

+++ EjerciciosUD6_SergioGonzalezCortes/src/ejercicio01/Calculadora.java
@@ -20,9 +20,9 @@ public class Calculadora {
 		return numero1 * numero2;
 	}
 	
-	public double dividir (int numero1, int numero2) {
+	public int dividir (int numero1, int numero2) {
 			
-		return (double)numero1 / (double)numero2;
+		return numero1 /numero2;
 	}
 	
 	public double calcularRaizCuadrada (int numero) {
